Handle default parameters explicitly
When an argument default value is gathered from a method definition be sure to pass also its associated argument when creating the Expression.Constant.

This avoids a problem which happens when the argument is null: in this case the type of the argument will be System.Object and that will fail when binding the arguments a method defining a different type (for example string)

In addition a new unit test was added to showcase this situation.